+++ Engine/CMakeLists.txt
@@ -419,6 +419,8 @@ target_sources(engine
     media/video/video.h
     platform/base/agsplatformdriver.cpp
     platform/base/agsplatformdriver.h
+    platform/base/agsplatform_unix.cpp
+    platform/base/agsplatform_unix.h
     platform/base/sys_main.cpp
     platform/base/sys_main.h
     platform/util/libc.c
