+++ Engine/platform/linux/acpllnx.cpp
@@ -16,38 +16,14 @@
 
 #if AGS_PLATFORM_OS_LINUX
 
-// ********* LINUX PLACEHOLDER DRIVER *********
+// *************** LINUX DRIVER ***************
 
-#include <stdio.h>
-#include <unistd.h>
-#include <allegro.h>
-#include "ac/runtime_defines.h"
-#include "gfx/gfxdefines.h"
 #include "platform/base/agsplatformdriver.h"
-#include "plugin/agsplugin.h"
-#include "util/string.h"
+#include "platform/base/agsplatform_unix.h"
 #include "libsrc/libcda-0.5/libcda.h"
 
-#include <pwd.h>
-#include <sys/stat.h>
-
-using AGS::Common::String;
-
-
-String CommonDataDirectory;
-String UserDataDirectory;
-
-struct AGSLinux : AGSPlatformDriver {
-
+struct AGSLinux : AGSPlatformUnix {
   int  CDPlayerCommand(int cmdd, int datt) override;
-  void DisplayAlert(const char*, ...) override;
-  const char *GetAllUsersDataDirectory() override;
-  const char *GetUserSavedgamesDirectory() override;
-  const char *GetUserConfigDirectory() override;
-  const char *GetUserGlobalConfigDirectory() override;
-  const char *GetAppOutputDirectory() override;
-  unsigned long GetDiskFreeSpaceMB() override;
-  const char* GetBackendFailUserHint() override;
   eScriptSystemOSID GetSystemOSID() override;
   int  InitializeCDPlayer() override;
   void ShutdownCDPlayer() override;
@@ -56,92 +32,6 @@ struct AGSLinux : AGSPlatformDriver {
 
 int AGSLinux::CDPlayerCommand(int cmdd, int datt) {
   return cd_player_control(cmdd, datt);
-}
-
-void AGSLinux::DisplayAlert(const char *text, ...) {
-  char displbuf[2000];
-  va_list ap;
-  va_start(ap, text);
-  vsprintf(displbuf, text, ap);
-  va_end(ap);
-  if (_logToStdErr)
-    fprintf(stderr, "%s\n", displbuf);
-  else
-    fprintf(stdout, "%s\n", displbuf);
-}
-
-size_t BuildXDGPath(char *destPath, size_t destSize)
-{
-  // Check to see if XDG_DATA_HOME is set in the enviroment
-  const char* home_dir = getenv("XDG_DATA_HOME");
-  size_t l = 0;
-  if (home_dir)
-  {
-    l = snprintf(destPath, destSize, "%s", home_dir);
-  }
-  else
-  {
-    // No evironment variable, so we fall back to home dir in /etc/passwd
-    struct passwd *p = getpwuid(getuid());
-    l = snprintf(destPath, destSize, "%s/.local", p->pw_dir);
-    if (mkdir(destPath, 0755) != 0 && errno != EEXIST)
-      return 0;
-    l += snprintf(destPath + l, destSize - l, "/share");
-    if (mkdir(destPath, 0755) != 0 && errno != EEXIST)
-      return 0;
-  }
-  return l;
-}
-
-void DetermineDataDirectories()
-{
-  if (!UserDataDirectory.IsEmpty())
-    return;
-  char xdg_path[256];
-  if (BuildXDGPath(xdg_path, sizeof(xdg_path)) == 0)
-    sprintf(xdg_path, "%s", "/tmp");
-  UserDataDirectory.Format("%s/ags", xdg_path);
-  mkdir(UserDataDirectory.GetCStr(), 0755);
-  CommonDataDirectory.Format("%s/ags-common", xdg_path);
-  mkdir(CommonDataDirectory.GetCStr(), 0755);
-}
-
-const char *AGSLinux::GetAllUsersDataDirectory()
-{
-  DetermineDataDirectories();
-  return CommonDataDirectory.GetCStr();
-}
-
-const char *AGSLinux::GetUserSavedgamesDirectory()
-{
-  DetermineDataDirectories();
-  return UserDataDirectory.GetCStr();
-}
-
-const char *AGSLinux::GetUserConfigDirectory()
-{
-  return GetUserSavedgamesDirectory();
-}
-
-const char *AGSLinux::GetUserGlobalConfigDirectory()
-{
-  return GetUserSavedgamesDirectory();
-}
-
-const char *AGSLinux::GetAppOutputDirectory()
-{
-  DetermineDataDirectories();
-  return UserDataDirectory.GetCStr();
-}
-
-unsigned long AGSLinux::GetDiskFreeSpaceMB() {
-  // placeholder
-  return 100;
-}
-
-const char* AGSLinux::GetBackendFailUserHint()
-{
-  return "Make sure you have latest version of SDL2 libraries installed, and X server is running.";
 }
 
 eScriptSystemOSID AGSLinux::GetSystemOSID() {
